render player positions fluently
