+++ web-app/src/static/js/main_script.js
@@ -1,31 +1,10 @@
 import { drawPitchOutline } from './render_pitch.js';
+import { resizeDrawable } from './drawable.js';
 
-const $ = (selector) => document.querySelector(selector);
-
-// Load control elements
-const controlsContainer = document.getElementById('controls-container');
-const tacticalboard = document.getElementById('tacticalboard');
 const drawable = document.getElementById('drawable');
-const toolbar = document.getElementById('toolbar');
+const tacticalboard = document.getElementById('tacticalboard');
+const controlsContainer = document.getElementById('controls-container');
 
-const drawable_ctx = drawable.getContext('2d');
-
-// set variables for drawable canvas
-let canvasOffsetX = drawable.offsetLeft;
-let canvasOffsetY = drawable.offsetTop;
-
-let style = getComputedStyle(drawable);
-drawable_ctx.canvas.width = parseInt(style.width);
-drawable_ctx.canvas.height = parseInt(style.height); 
-
-let isPainting = false;
-let lineWidth = 5;
-let startX;
-let startY;
-
-//---------------------------------------------
-// FUNCTIONS
-//---------------------------------------------
 
 const resizeTactical = () => {
     const controlsContainerHeight = controlsContainer.offsetHeight;
@@ -40,15 +19,6 @@ const resizeTactical = () => {
     drawable.width = controlsWidth;
     drawable.height = controlsHeight;
   }
-
-const resizeDrawable = () => { 
-    style = getComputedStyle(drawable);
-    drawable_ctx.canvas.width = parseInt(style.width);
-    drawable_ctx.canvas.height = parseInt(style.height);
-
-    canvasOffsetX = drawable.offsetLeft;
-    canvasOffsetY = drawable.offsetTop;
-} 
 
 //---------------------------------------------
 // EVENT LISTENERS
@@ -67,49 +37,3 @@ window.addEventListener('resize', e => {
     resizeDrawable();
     resizeDrawable();
 });
-
-//---------------------------------------------
-// Drawable Event Listeners
-//---------------------------------------------
-toolbar.addEventListener('click', e => {
-    if (e.target.id === 'clear') {
-        drawable_ctx.clearRect(0, 0, drawable.width, drawable.height);
-    }
-});
-
-toolbar.addEventListener('change', e => {
-    if(e.target.id === 'stroke') {
-        drawable_ctx.strokeStyle = e.target.value;
-    }
-
-    if(e.target.id === 'lineWidth') {
-        lineWidth = e.target.value;
-    }
-    
-});
-
-const draw = (e) => {
-    if(!isPainting) {
-        return;
-    }
-
-    drawable_ctx.lineWidth = lineWidth;
-    drawable_ctx.lineCap = 'round';
-
-    drawable_ctx.lineTo(e.clientX - canvasOffsetX, e.clientY - canvasOffsetY);
-    drawable_ctx.stroke();
-}
-
-drawable.addEventListener('mousedown', (e) => {
-    isPainting = true;
-    startX = e.clientX;
-    startY = e.clientY;
-});
-
-drawable.addEventListener('mouseup', e => {
-    isPainting = false;
-    drawable_ctx.stroke();
-    drawable_ctx.beginPath();
-});
-
-drawable.addEventListener('mousemove', draw);
