+++ web-app/src/templates/index.html
@@ -15,7 +15,7 @@
 
     <!-- import the webpage's javascript file -->
     <script type="module"  src="{{url_for('static', filename='js/video_analysis.js')}}" defer></script>
-    <!-- <script src="{{url_for('static', filename='js/drawable.js')}}" defer></script> -->
+    <script type="module" src="{{url_for('static', filename='js/drawable.js')}}" defer></script>
     <script type="module" src="{{url_for('static', filename='js/main_script.js')}}" defer></script>
     <script type="module" src="{{url_for('static', filename='js/render_pitch.js')}}" defer></script>
